test: add markdown parser unit tests
