+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc -b && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "dependencies": {
     "dayjs": "^1.11.13",
@@ -32,6 +33,7 @@
     "sass-embedded": "^1.86.3",
     "typescript": "~5.7.2",
     "vite": "^6.2.0",
+    "vitest": "^3.2.2",
     "vue-tsc": "^2.2.4"
   }
 }
