+++ vite.config.ts
@@ -10,4 +10,8 @@ export default defineConfig({
       "@": fileURLToPath(new URL("./src", import.meta.url)),
     },
   },
+  test: {
+    globals: true,
+    environment: 'node',
+  },
 })
